% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllowableBearingCapacity_qa.R
\name{AllowableBearingCapacity_qa}
\alias{AllowableBearingCapacity_qa}
\title{Title}
\usage{
AllowableBearingCapacity_qa(qu, FS)
}
\arguments{
\item{qu}{Ultimate Bearing Capacity Pressure, qu, from Buisman-Terzaghi Model}

\item{FS}{Bench Factor of Safety(between 2 and 4)}
}
\description{
Title
}
