% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvStressFoundation_qd.R
\name{AvStressFoundation_qd}
\alias{AvStressFoundation_qd}
\title{Title Calculate the Average Stress Acting on the Foundation (qd);}
\usage{
AvStressFoundation_qd(Q, B, L)
}
\arguments{
\item{Q}{The load on each track (Q)}

\item{B}{track width}

\item{L}{Track Contact Length}
}
\description{
Title Calculate the Average Stress Acting on the Foundation (qd);
}
