% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_Blastability.R
\name{Berg_Blastability}
\alias{Berg_Blastability}
\title{Title Blastability Index according to Lilly, (1986)}
\usage{
Berg_Blastability(RMD, JPS, JPA, RDI, HF)
}
\arguments{
\item{RMD}{Rock mass description}

\item{JPS}{Joint Plane Spacing}

\item{JPA}{Joint Plane Angle}

\item{RDI}{Rock Density Influence}

\item{HF}{Hardness Factor}
}
\description{
Title Blastability Index according to Lilly, (1986)
}
