% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_KCO_50.R
\name{Berg_KCO_50}
\alias{Berg_KCO_50}
\title{Title  Mean Fragment Estimation from KCO Model}
\usage{
Berg_KCO_50(RF, PF, QE, RWS)
}
\arguments{
\item{RF}{Rock Factor}

\item{PF}{Powder Factor}

\item{QE}{Quantity/amount of explosives}

\item{RWS}{Relative weight strength (RWS = weight strength relative to ANFO )}
}
\description{
Title  Mean Fragment Estimation from KCO Model
}
