% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_KCO_Fragmentation.R
\name{Berg_KCO_Fragmentation}
\alias{Berg_KCO_Fragmentation}
\title{Title Fragment size distribution}
\usage{
Berg_KCO_Fragmentation(
  Data_Size = data.frame(Data_Size = c(0, 0.2, 0.475, 0.635, 0.953, 1.27, 1.905, 2.54,
    5.08, 10.16, 15.24, 20.32, 25.4, 38.1, 63.5, 127, 250, 350), stringsAsFactors =
    FALSE),
  KCO_50,
  KCO_BUND,
  XMAX
)
}
\arguments{
\item{Data_Size}{Size Fraction}

\item{KCO_50}{Mean fragment size  computed from KCO Model}

\item{KCO_BUND}{itle Curver Undulation factor/parameter (b)}

\item{XMAX}{min(in situ block size; S or B)}
}
\description{
Title Fragment size distribution
}
