% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_KUZRM_50.R
\name{Berg_KUZRM_50}
\alias{Berg_KUZRM_50}
\title{Title Mean Fragmentation Size estimation according to  Cunningham(2005)}
\usage{
Berg_KUZRM_50(RF, PF, QE, RWS)
}
\arguments{
\item{RF}{Rock factor}

\item{PF}{Powder factor}

\item{QE}{mass of explosive in the hole, kg or quantity/amount of explosive per hole}

\item{RWS}{weight strength relative to ANFO}
}
\description{
Title Mean Fragmentation Size estimation according to  Cunningham(2005)
}
