% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_Kuzrm_Fragmentation.R
\name{Berg_Kuzrm_Fragmentation}
\alias{Berg_Kuzrm_Fragmentation}
\title{Title Fragment size Distribution according to Cunningham(2005)}
\usage{
Berg_Kuzrm_Fragmentation(
  Data_Size = data.frame(Data_Size = c(0, 0.2, 0.475, 0.635, 0.953, 1.27, 1.905, 2.54,
    5.08, 10.16, 15.24, 20.32, 25.4, 38.1, 63.5, 127, 250, 350), stringsAsFactors =
    FALSE),
  KUZRM_50,
  KZRM_n,
  ...
)
}
\arguments{
\item{Data_Size}{Size Fraction}

\item{KUZRM_50}{Mean Fragment size according to Cunningham(2005)}

\item{KZRM_n}{Uniformity index or exponent according to Cunningham(2005)}
}
\description{
Title Fragment size Distribution according to Cunningham(2005)
}
