% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_KZRM_n.R
\name{Berg_KZRM_n}
\alias{Berg_KZRM_n}
\title{Title Uniformity Index/exponent calculation for staggered and squared pattern according to Cunningham(2005)}
\usage{
Berg_KZRM_n(Pattern_Blast, BCL, CCL, B, d, S, W, CL, BH, ...)
}
\arguments{
\item{Pattern_Blast}{}

\item{BCL}{the bottom charge length, m;}

\item{CCL}{the column charge length, m}

\item{B}{burden, m;}

\item{d}{the hole diameter, mm}

\item{S}{spacing, m;}

\item{W}{the standard deviation of drilling precision, m;}

\item{CL}{is the charge length, m;}

\item{BH}{the bench height, m.}

\item{...}{}
}
\description{
Title Uniformity Index/exponent calculation for staggered and squared pattern according to Cunningham(2005)
}
\examples{
Berg_KZRM_n( Pattern_Blast = "Staggered Pattern", BCL = 0, CCL = 0, B= 5, d = 235, S = 5.5, W = 0.15, CL = 3.85, BH = 6     )
}
