% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_MKUZRM_50.R
\name{Berg_MKUZRM_50}
\alias{Berg_MKUZRM_50}
\title{Title Estimation of Modified Kuz-Ram Model according to (Gheibie et al., 2009).}
\usage{
Berg_MKUZRM_50(BI, VB, QE, SANFO, ...)
}
\arguments{
\item{BI}{the blastability index}

\item{VB}{the volume of rock broken by one blasthole}

\item{QE}{mass of explosive in each hole, kg}

\item{SANFO}{the relative weight strength of the explosive to ANFO;}
}
\description{
Title Estimation of Modified Kuz-Ram Model according to (Gheibie et al., 2009).
}
