% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_MKuzrm_Fragmentation.R
\name{Berg_MKuzrm_Fragmentation}
\alias{Berg_MKuzrm_Fragmentation}
\title{Title Fragment Size Distribution according to Modified Kuz-Ram Model}
\usage{
Berg_MKuzrm_Fragmentation(
  Data_Size = data.frame(Data_Size = c(0, 0.2, 0.475, 0.635, 0.953, 1.27, 1.905, 2.54,
    5.08, 10.16, 15.24, 20.32, 25.4, 38.1, 63.5, 127, 250, 350), stringsAsFactors =
    FALSE),
  MKUZRM_50,
  MKZRM_n,
  ...
)
}
\arguments{
\item{Data_Size}{Size Fraction}

\item{MKUZRM_50}{the mean fragment size, cm}

\item{MKZRM_n}{is the modified uniformity index}

\item{...}{}
}
\description{
Title Fragment Size Distribution according to Modified Kuz-Ram Model
}
