% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_Modified_n.R
\name{Berg_Modified_n}
\alias{Berg_Modified_n}
\title{Title modified uniformity index calculation according to (Gheibie et al., 2009).}
\usage{
Berg_Modified_n(RMD, JPS, JPA, RDI, HF, Berg_KZRM_n, ...)
}
\arguments{
\item{RMD}{Rock mass description}

\item{JPS}{Joint Plane Spacing}

\item{JPA}{Joint Plane Angle}

\item{RDI}{Rock Density Influence}

\item{HF}{Hardness Factor}

\item{Berg_KZRM_n}{the uniformity index ( Cunningham’s uniformity index , see Cunningham(2005))}

\item{...}{}
}
\description{
Title modified uniformity index calculation according to (Gheibie et al., 2009).
}
