% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_PF.R
\name{Berg_PF}
\alias{Berg_PF}
\title{Title Estimation of Powder factor}
\usage{
Berg_PF(B, S, H, QE, ...)
}
\arguments{
\item{B}{Burden}

\item{S}{Spacing}

\item{H}{Bench height}

\item{QE}{mass of explosive in the hole, kg; /qauntity explosive per hole}
}
\description{
Title Estimation of Powder factor
}
