% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_RF.R
\name{Berg_RF}
\alias{Berg_RF}
\title{Title Estimation of Rock Factor}
\usage{
Berg_RF(RMD, RDI, HF, ...)
}
\arguments{
\item{RMD}{Rock mass description}

\item{RDI}{Rock Density Influence}

\item{HF}{Hardness Factor}

\item{...}{}
}
\description{
Title Estimation of Rock Factor
}
