% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berg_Volume.R
\name{Berg_Volume}
\alias{Berg_Volume}
\title{Title Esimation of Volume (Bank Cubic Meter, BCM)}
\usage{
Berg_Volume(B, S, H, ...)
}
\arguments{
\item{B}{Burden}

\item{S}{Spacing}

\item{H}{Bench Heigh}
}
\description{
Title Esimation of Volume (Bank Cubic Meter, BCM)
}
