% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bergblast_HF.R
\name{Bergblast_HF}
\alias{Bergblast_HF}
\title{Title Estimation of Hardness Factor}
\usage{
Bergblast_HF(YM, UCS, ...)
}
\arguments{
\item{YM}{Young's modulus}

\item{UCS}{Uniaxial Compressive Strength (UCS) in MPa}
}
\description{
Title Estimation of Hardness Factor
}
