% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchBench_NOSwell.R
\name{CatchBench_NOSwell}
\alias{CatchBench_NOSwell}
\title{Title Catch Bench Width without swell factor condition}
\usage{
CatchBench_NOSwell(H, BehtaW, Behta, r, SBW)
}
\arguments{
\item{H}{Bench Height}

\item{BehtaW}{Dip of plunge of failure plane}

\item{Behta}{Bench Face Angle}

\item{r}{Angle of repose}

\item{SBW}{safety berm Width or The catch bench width}
}
\description{
Title Catch Bench Width without swell factor condition
}
\examples{
CatchBench_NOSwell( H = 50 , Behta= 65, BehtaW = 45, r = 30 , SBW = 12   ) 
}
