% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatchBench_Swell.R
\name{CatchBench_Swell}
\alias{CatchBench_Swell}
\title{Title Catch Bench Width swell factor condition}
\usage{
CatchBench_Swell(H, BehtaW, Behta, r, SBW, SF)
}
\arguments{
\item{H}{Bench Height
@param SF Swell Factor}

\item{BehtaW}{Dip of plunge of failure plane}

\item{Behta}{Bench Face Angle}

\item{r}{Angle of repose}

\item{SBW}{safety berm Width or The catch bench width}
}
\description{
Title Catch Bench Width swell factor condition
}
\examples{
CatchBench_Swell(  H = 50 , Behta= 65, BehtaW = 45, r = 30 , SBW = 12 , SF= 1.3   ) #H = 50 ft , B = 65, Bw = 45, r = 30 , SBw = 12 , SF= 1.3
}
