% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chatPost}
\alias{chatPost}
\title{Post to chat}
\usage{
chatPost(text)
}
\arguments{
\item{text}{A character string with the message you would like
to post.}
}
\description{
Post a message to Minecraft chat.
}
\examples{
\dontrun{
chatPost('foobar')

## Post the first few digits of pi to chat
for (p in strsplit(as.character(round(pi, 4)), '')[[1]]) {
    chatPost(p)
  }
}

}
