% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1ShearNormalStress.R
\name{F1ShearNormalStress}
\alias{F1ShearNormalStress}
\title{Title}
\usage{
F1ShearNormalStress(c, A, W, psi, phi)
}
\arguments{
\item{c}{Cohesion}

\item{A}{base area of block}

\item{W}{weight of block}

\item{psi}{Slope face angle}

\item{phi}{Friction angle}
}
\description{
Title
}
