% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F2EffectsOfWaterPressure.R
\name{F2EffectsOfWaterPressure}
\alias{F2EffectsOfWaterPressure}
\title{Title}
\usage{
F2EffectsOfWaterPressure(c, A, W, psi, phi, u)
}
\arguments{
\item{c}{cohesion}

\item{A}{base area of block}

\item{W}{weight of block}

\item{psi}{Slope face angle}

\item{phi}{Friction angle}

\item{u}{Water Pressure}
}
\description{
Title
}
