% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F3EffectsWatTenCrackPressure.R
\name{F3EffectsWatTenCrackPressure}
\alias{F3EffectsWatTenCrackPressure}
\title{Title}
\usage{
F3EffectsWatTenCrackPressure(c, A, W, psi, phi, U, V)
}
\arguments{
\item{c}{cohesion}

\item{A}{base area of block}

\item{W}{weight of block}

\item{psi}{Slope face angle}

\item{phi}{Friction angle}

\item{U}{Water Pressure}

\item{V}{Pressure,water in the tension crack}
}
\description{
Title
}
