% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F4EffectReinforcementRbolts.R
\name{F4EffectReinforcementRbolts}
\alias{F4EffectReinforcementRbolts}
\title{Title}
\usage{
F4EffectReinforcementRbolts(c, A, W, psi, phi, U, V, Tee, Behta, ...)
}
\arguments{
\item{c}{cohesion}

\item{A}{base area of block}

\item{W}{weight of block}

\item{psi}{Slope face angle}

\item{phi}{Friction angle}

\item{U}{Water Pressure}

\item{V}{Pressure,water in the tension crack}

\item{Tee}{Force Provided by the Rockbolts}

\item{Behta}{rock bolt angle to the failure plane}

\item{...}{}
}
\description{
Title
}
