% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_entity.R
\name{find_entity}
\alias{find_entity}
\title{Find entity by name}
\usage{
find_entity(name)
}
\arguments{
\item{name}{Character string with a Minecraft entity name (can be a partial match)}
}
\value{
Data frame with a row or set of rows from the output of
\code{\link[=getEntityTypes]{getEntityTypes()}} that match the queried name.
}
\description{
Find a Minecraft entity by name.
}
\details{
We first look to see whether there is an exact match to
the \code{namex} column in the output of \code{\link[=getEntityTypes]{getEntityTypes()}}. If there
is, we return that row. If not, we use \code{\link[=grep]{grep()}} with
\code{ignore.case=TRUE} and return matching rows.
}
\examples{
find_entity(name = "witch")

}
\seealso{
\code{\link[=getEntityTypes]{getEntityTypes()}}, \code{\link[=find_items]{find_items()}}
}
