% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlock.R
\name{getBlock}
\alias{getBlock}
\title{Determine block type at some position}
\usage{
getBlock(x, y, z)
}
\arguments{
\item{x}{A numeric string with north/south position}

\item{y}{A numeric string with height}

\item{z}{A numeric string with east/west position}
}
\value{
A character string with the ID of the block type at the
   position (x, y, z). You can use \code{\link{find_item}} to 
   find the name of the block type based on this returned ID.
}
\description{
Determine the type of the block at position x (north / south),
y (height), z (east / west).
}
\examples{
\dontrun{
mc_connect()
h <- getHeight(0,0)
b_type <- getBlock(0,h,0)
b_type

find_item(id = b_type)
}

}
