% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{getBlockHits}
\alias{getBlockHits}
\title{Get most recent block hits}
\usage{
getBlockHits()
}
\value{
A dataframe with columns for the coordinates, block type,
    and player id of recent block hits.
}
\description{
Return the most recent block hits made in the Minecraft
world by an iron sword.
}
\note{
Only right clicks with an iron sword are logged.
}
\examples{
\dontrun{
getBlockHits()

library(futile.logger)
while (TRUE) {
    flog.info(getEventsBlockHits())
    Sys.sleep(1)
}
}

}
\seealso{
\code{\link{getChatPosts}}
}
