% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlock.R
\name{getBlockWithData}
\alias{getBlockWithData}
\title{Determine block type and style at some position}
\usage{
getBlockWithData(con, x, y, z)
}
\arguments{
\item{con}{Socket connection to minecraft server}

\item{x}{north/south position}

\item{y}{height}

\item{z}{east/west position}
}
\value{
None.
}
\description{
Determine block type and style at position (x,y,z)
}
\examples{
\dontrun{
mc <- mc_connect()
h <- getHeight(mc, 0,0)
getBlockWithStyle(mc,  0,h,0)
}

}
