% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{getChatPosts}
\alias{getChatPosts}
\title{Pull the most recent chat message}
\usage{
getChatPosts()
}
\value{
A dataframe of recent chat posts with columns for the player 
   id of the poster and the message posted
}
\description{
Returns the chat posts that have been posted recently and since the
last time the function was run. It is possible that this does not 
return chat posts that were posted using \code{\link{chatPost}}.
}
\examples{
\dontrun{
mc_connect()
getChatPosts()
}

}
\seealso{
\code{\link{getBlockHits}}
}
