% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerOrientation.R
\name{getDirection}
\alias{getDirection}
\title{Get direction vector}
\usage{
getDirection()
}
\value{
A vector of length 3 with coordinates of current direction
}
\description{
Get direction vector
}
\examples{
\dontrun{
getDirection()
}
}
