% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{getEventsBlockHits}
\alias{getEventsBlockHits}
\title{Read most recent block hits}
\usage{
getEventsBlockHits()
}
\value{
\code{data.frame} of coordinates, block type and user id
}
\description{
Note, that only right clicks with an Iron sword are logged.
}
\examples{
\dontrun{
library(futile.logger)
while (TRUE) {
    flog.info(getEventsBlockHits())
    Sys.sleep(1)
}
}
}
\seealso{
getEventsChatPosts
}
