% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{getEventsChatPosts}
\alias{getEventsChatPosts}
\title{Read most recent chat messages}
\usage{
getEventsChatPosts()
}
\value{
\code{data.frame} player id and message
}
\description{
Read most recent chat messages
}
\seealso{
getEventsBlockHits
}
