% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{getHeading}
\alias{getHeading}
\title{Get player heading}
\usage{
getHeading(player_id)
}
\arguments{
\item{player_id}{An integer with the player id}
}
\value{
Player heading in degrees
}
\description{
Get player heading
}
\examples{
\dontrun{
getHeading(myid)
}

}
