% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeight.R
\name{getHeight}
\alias{getHeight}
\title{Get height of the world at a given position}
\usage{
getHeight(x, z)
}
\arguments{
\item{x}{A numeric string with north/south position}

\item{z}{A numeric string with east/west position}
}
\value{
Integer with height of block just above the last bit of
non-air at the specified x / z position.
}
\description{
Get height of the world at a given position, where the world's
height is defined as the height of the highest non-air point at
that location.
}
\examples{
\dontrun{
mc_connect()
getHeight(0,0)
}

}
