% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerOrientation.R
\name{getPitch}
\alias{getPitch}
\title{Get player pitch}
\usage{
getPitch()
}
\value{
a vector double of length one with pitch
}
\description{
Return player pitch
}
\examples{
\dontrun{
getPitch()
}
}
