% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerId.R
\name{getPlayerId}
\alias{getPlayerId}
\title{Get player ID}
\usage{
getPlayerId(player_name)
}
\arguments{
\item{player_name}{Character string with name of a player.}
}
\value{
A number representing the ID of the player.
}
\description{
Get the ID for a player with a given name.
}
\examples{
\dontrun{
mc_connect()
getPlayerIds()
getPlayerId("wyldstyle")
}

}
\seealso{
\code{\link[=getPlayerIds]{getPlayerIds()}}, \code{\link[=getPlayerName]{getPlayerName()}}
}
