% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerIds.R
\name{getPlayerIds}
\alias{getPlayerIds}
\title{Get player IDs}
\usage{
getPlayerIds(con)
}
\arguments{
\item{con}{Socket connection to minecraft server}
}
\value{
set of integers
}
\description{
Get player IDs
}
\examples{
\dontrun{
mc <- mc_connect()
p <- getPlayerPost(mc)
setPlayerPos(mc, 0, p + 5, 0)
}

}
