% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerName.R
\name{getPlayerName}
\alias{getPlayerName}
\title{Get player name}
\usage{
getPlayerName(player_id)
}
\arguments{
\item{player_id}{Integer giving the ID of a player. You can find IDs of all
current players using \code{\link[=getPlayerIds]{getPlayerIds()}}.}
}
\value{
A character string with the name of the identified player.
}
\description{
Get the name of a player with a given ID
}
\examples{
\dontrun{
mc_connect()
getPlayerIds()
getPlayerName(322)
}

}
\seealso{
\code{\link[=getPlayerIds]{getPlayerIds()}}, \code{\link[=getPlayerId]{getPlayerId()}}
}
