% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerOrientation.R
\name{getPlayerPitch}
\alias{getPlayerPitch}
\title{Get player pitch}
\usage{
getPlayerPitch()
}
\value{
a vector double of length one with pitch
}
\description{
Return player pitch
}
\examples{
\dontrun{
getPitch()
}
}
