% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerPos.R
\name{getPlayerPos}
\alias{getPlayerPos}
\title{Get player position}
\usage{
getPlayerPos()
}
\value{
vector (x,y,z)
}
\description{
Get player position
}
\examples{
\dontrun{
mc <- mc_connect()
getPlayerPos()
}

}
\seealso{
[getPlayerPos]
}
