% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerOrientation.R
\name{getPlayerRotation}
\alias{getPlayerRotation}
\title{Get player rotation}
\usage{
getPlayerRotation()
}
\value{
double of length 1 with rotatio
}
\description{
Get player rotation from current position
}
\examples{
\dontrun{
getPlayerRotation()
}

}
\seealso{
[getPlayerDirection]
}
