% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerPos.R
\name{getPlayerTilePos}
\alias{getPlayerTilePos}
\title{Get player tile position}
\usage{
getPlayerTilePos(con)
}
\arguments{
\item{con}{Socket connection to minecraft server}
}
\value{
vector (x,y,z)
}
\description{
Get player tile position
}
\examples{
\dontrun{
mc <- mc_connect()
getPlayerTilePos(mc)
}

}
\seealso{
[getPlayerPos]
}
