% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{initHeading}
\alias{initHeading}
\title{Set player heading to match the direction the player is facing}
\usage{
initHeading(player_id)
}
\arguments{
\item{player_id}{An integer with the player id}
}
\description{
Set player heading to match the direction the player is facing
}
\examples{
\dontrun{
initHeading(myid)
}

}
