% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCO_Bund.R
\name{KCO_Bund}
\alias{KCO_Bund}
\title{KCO_Bund: Title Curver Undulation factor/parameter (b) 	        F. Ouchterlony and J.A. Sanchidrián, 2019  https://doi.org/10.1016/j.jrmge.2019.03.001}
\usage{
KCO_Bund(KCO_50, Berg_KZRM_n, XMAX)
}
\arguments{
\item{KCO_50}{Mean Fragment size computed from the KCO Model}

\item{Berg_KZRM_n}{}

\item{XMAX}{min(in situ block size; S or B)}
}
\description{
KCO_Bund: Title Curver Undulation factor/parameter (b) 	        F. Ouchterlony and J.A. Sanchidrián, 2019  https://doi.org/10.1016/j.jrmge.2019.03.001
}
