% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{lookForward}
\alias{lookForward}
\title{Check the block type (if any) ahead of the player}
\usage{
lookForward(player_id, distance = 1)
}
\arguments{
\item{player_id}{An integer with the player id}

\item{distance}{The distance to look}
}
\description{
Check the block type (if any) ahead of the player.
Look in the direction of the player heading
at the specified distance.
}
\examples{
\dontrun{
lookForward(myid)
}

}
