% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{mc_close}
\alias{mc_close}
\title{Close cached connection to Minecraft server}
\usage{
mc_close()
}
\value{
None.
}
\description{
Close the current connection to a Minecraft server.
}
\examples{
\dontrun{
  mc_connect()
  getPlayerIds()
  mc_close()
}

}
