% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maze.R
\name{mc_maze}
\alias{mc_maze}
\title{Generates a maze in front of a player}
\usage{
mc_maze(n = 5, player_id = NULL)
}
\arguments{
\item{n}{size of the maze}

\item{player_id}{optional player id}
}
\description{
Generates a maze in front of a player
}
