% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maze.R
\name{mc_mazer}
\alias{mc_mazer}
\title{Poll the chat window for maze generator commands and spawn mazes right in front of the player}
\usage{
mc_mazer()
}
\description{
Poll the chat window for maze generator commands and spawn mazes right in front of the player
}
