% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{mcsplit_message}
\alias{mcsplit_message}
\title{Helper function to split string into data.frame, for the chat}
\usage{
mcsplit_message(string, colnames)
}
\description{
Helper function to split string into data.frame, for the chat
}
\keyword{internal}
