% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miner-package.R
\docType{package}
\name{miner-package}
\alias{miner}
\alias{miner-package}
\title{miner: Control Minecraft with RaspberryJuice API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Following the code in py3minepi, allows one to connect to the minecraft from R.
}
\section{Vignette}{

\itemize{
\item \href{https://kbroman.org/miner/}{Program in Minecraft with R and miner}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/miner}
  \item Report bugs at \url{https://github.com/kbroman/miner/issues}
}

}
\author{
\strong{Maintainer}: Karl Broman \email{broman@wisc.edu}

Authors:
\itemize{
  \item Karl Broman \email{broman@wisc.edu}
  \item Brooke Anderson
  \item Gergely Daróczi
  \item Ali Zaidi
  \item Mario Inchiosa
  \item David Smith
}

}
\keyword{internal}
