% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{moveForward}
\alias{moveForward}
\title{Move the player in the direction of the player heading}
\usage{
moveForward(player_id, distance = 1)
}
\arguments{
\item{player_id}{An integer with the player id}

\item{distance}{The distance to move the player}
}
\description{
Move the player in the direction of the player heading
}
\examples{
\dontrun{
initHeading(myid)
moveForward(myid, 0.5)
}

}
