% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Num_TracksBWE.R
\name{Num_TracksBWE}
\alias{Num_TracksBWE}
\title{Title}
\usage{
Num_TracksBWE(B, L, qu, LoadGround)
}
\arguments{
\item{B}{track width(B)}

\item{L}{Track Contact Length}

\item{qu}{Ultimate Bearing Capacity}

\item{LoadGround}{total load to betransferred to the ground}
}
\description{
Title
}
\examples{
Num_TracksBWE( B = 3, L = 14, qu = 1885.78,LoadGround = 225000   )
}
