% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall_Pit_Slope.R
\name{Overall_Pit_Slope}
\alias{Overall_Pit_Slope}
\title{Title The overall pit slope, is given by Theta}
\usage{
Overall_Pit_Slope(n, H, BW, alphadeg)
}
\arguments{
\item{n}{number of benches}

\item{H}{bench height}

\item{BW}{bench width}

\item{alphadeg}{bench slope}
}
\description{
Title The overall pit slope, is given by Theta
}
\examples{
 Overall_Pit_Slope(n = 5, H = 50,  BW = 35, alphadeg = 75)  #n = 5; H = 50 ft ; Bw = b = 35 ft; alphadeg  = 75
}
