% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall_Pit_Slope_Effect_of_Ramp.R
\name{Overall_Pit_Slope_Effect_of_Ramp}
\alias{Overall_Pit_Slope_Effect_of_Ramp}
\title{Title The overall pit slope,Effect of Ramps}
\usage{
Overall_Pit_Slope_Effect_of_Ramp(n, H, BW, alphadeg, RW)
}
\arguments{
\item{n}{number of benches}

\item{H}{bench height}

\item{BW}{bench width}

\item{alphadeg}{bench slope}

\item{RW}{Ramp Width}
}
\description{
Title The overall pit slope,Effect of Ramps
}
\examples{
 Overall_Pit_Slope_Effect_of_Ramp(n = 5, H = 50,  BW = 35, alphadeg = 75, RW = 100)  #n = 5; H = 50 ft ; Bw = b = 35 ft; alphadeg  = 75, RW  = 100ft
}
