% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall_Pit_Slope_EffectofInt_Ramps.R
\name{Overall_Pit_Slope_EffectofInt_Ramps}
\alias{Overall_Pit_Slope_EffectofInt_Ramps}
\title{Title The overall pit slope, Effect of Inter-Ramps}
\usage{
Overall_Pit_Slope_EffectofInt_Ramps(n, H, BW, alphadeg, RW)
}
\arguments{
\item{n}{number of benches}

\item{H}{bench height}

\item{BW}{bench width}

\item{alphadeg}{bench slope}
}
\description{
Title The overall pit slope, Effect of Inter-Ramps
}
\examples{
 Overall_Pit_Slope_EffectofInt_Ramps(n = 5, H = 50,  BW = 35, alphadeg = 75)  #n = 5; H = 50 ft ; Bw = b = 35 ft; alphadeg  = 75
}
