% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overall_Pit_Slope_EffectofMultWorkingBenches.R
\name{Overall_Pit_Slope_EffectofMultWorkingBenches}
\alias{Overall_Pit_Slope_EffectofMultWorkingBenches}
\title{Title The overall pit slope, Effects of Multiple Working Benches}
\usage{
Overall_Pit_Slope_EffectofMultWorkingBenches(
  n,
  n1,
  n2,
  H,
  BW,
  alphadeg,
  WBW,
  ...
)
}
\arguments{
\item{n1}{number of benches Two working benches in the pit layout ( n1 # of upper benches , n2 # of lower benches}

\item{H}{bench height}

\item{BW}{bench width}

\item{alphadeg}{bench slope}

\item{WBW}{working benches}
}
\description{
Title The overall pit slope, Effects of Multiple Working Benches
}
\examples{
 Overall_Pit_Slope_EffectofMultWorkingBenches(n = 6, H = 50,  BW = 35, alphadeg = 75,   n1 = 3, n2 = 2, WBW = 125)  #n = 5; H = 50 ft ; Bw = b = 35 ft; alphadeg  = 75,WBW = 125ft,  n1 = 3, n2 = 2 
}
