% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlock.R
\name{setBlock}
\alias{setBlock}
\title{Place a block}
\usage{
setBlock(con, x, y, z, block_id, block_style = 0)
}
\arguments{
\item{con}{Socket connection to minecraft server}

\item{x}{north/south position}

\item{y}{height}

\item{z}{east/west position}

\item{block_id}{block id}

\item{block_style}{block_style}
}
\value{
None.
}
\description{
Place a block at position (x,y,z) by type id
}
\examples{
\dontrun{
mc <- mc_connect()
h <- getHeight(mc, 0,0)
setBlock(mc,  0,h,0,  46)
}

}
