% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{setHeading}
\alias{setHeading}
\title{Set player heading}
\usage{
setHeading(player_id, new_heading)
}
\arguments{
\item{player_id}{An integer with the player id}

\item{new_heading}{A numeric with the direction to head forward}
}
\description{
Set the direction in which moveForward will move the player.
The heading is measured in degrees.
A heading of 180 or -180 degreees is towards North or negative z
A heading of 0 degreees is towards South or positive z
A heading of 270 or -90 degreees is towards East or positive x
A heading of 90 degrees is towards West or negative x
}
\examples{
\dontrun{
setHeading(myid, 0)
}

}
