% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlayerPos.R
\name{setPlayerPos}
\alias{setPlayerPos}
\title{Change player position}
\usage{
setPlayerPos(x, y, z, player_id = NULL, tile = FALSE)
}
\arguments{
\item{x}{A numeric string with north/south position}

\item{y}{A numeric string with height}

\item{z}{A numeric string with east/west position}

\item{player_id}{Integer giving the ID of a player. You can find IDs of all
current players using \code{\link[=getPlayerIds]{getPlayerIds()}}.}

\item{tile}{Logical value specifying whether to truncate the output position to
an integer (i.e., the location of the tile on which the player is
positioned).}
}
\value{
None.
}
\description{
Move player to position (x,y,z). The default is to move the first player who
was spawned in the Minecraft world, but other players can also be moved using
the \code{player_id} argument. If the \code{tile} argument is set to \code{TRUE},
the player will be moved to the position specified by truncating the specified
x, y, and z to integers.
}
\examples{
\dontrun{
mc_connect()
p <- getPlayerPos()
setPlayerPos(0, p + 5, 0)

example_entity <- getPlayerIds()[1]
getPlayerPos(example_entity)
setPlayerPos(0, p, 0, example_entity)
getPlayerPos(example_entity)
}

}
