% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlayerPos.R
\name{setPlayerTilePos}
\alias{setPlayerTilePos}
\title{Change player position to a tile}
\usage{
setPlayerTilePos(con, x, y, z)
}
\arguments{
\item{con}{Socket connection to minecraft server}

\item{x}{north/south position}

\item{y}{height}

\item{z}{east/west position}
}
\value{
None.
}
\description{
Move player to tile position (x,y,z)
}
\examples{
\dontrun{
mc <- mc_connect()
p <- getPlayerPost(mc)
setPlayerPos(mc, 0, p + 5, 0)
}

}
