% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soilweight_q.R
\name{Soilweight_q}
\alias{Soilweight_q}
\title{Title}
\usage{
Soilweight_q(gamma_prime, D)
}
\arguments{
\item{gamma_prime}{effective soil bulk density}

\item{D}{depth of track penetration(D )}
}
\description{
Title
}
\examples{
Soilweight_q(gamma_prime = 1.8 , D =0.1 )  #answer in t/m3
}
