% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{turnLeft}
\alias{turnLeft}
\title{Rotate the player heading to the left}
\usage{
turnLeft(player_id, angle = 90)
}
\arguments{
\item{player_id}{An integer with the player id}

\item{angle}{The number of degrees to rotate left}
}
\description{
Use a positive angle value to rotate to the left.
Use a negative angle to rotate to the right.
The player does not turn in the game; rather, the heading
for subsequent movement commands is rotated.
}
\examples{
\dontrun{
turnLeft(myid)
}

}
