% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UltBearingCapPressure_qu.R
\name{UltBearingCapPressure_qu}
\alias{UltBearingCapPressure_qu}
\title{Title}
\usage{
UltBearingCapPressure_qu(c, Nec, D, q, Neq, gamma_prime, B, N_gamma, phi)
}
\arguments{
\item{c}{soil cohesion}

\item{Nec}{Cohesion Factor}

\item{D}{depth of track penetration}

\item{q}{soil weight}

\item{Neq}{Soil Weight Factor}

\item{gamma_prime}{effective soil bulk density}

\item{B}{track width}

\item{N_gamma}{Machine-Soil Factor}

\item{phi}{angle of internal friction for soil}
}
\description{
Title
}
