% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkingBench_Width.R
\name{WorkingBench_Width}
\alias{WorkingBench_Width}
\title{Title Working Bench Design
SBW = 2*HSB , HSB  = Height of Safety Berm, HSB, Greater than or equal to (>=)  Height of Largest Truck Axle}
\usage{
WorkingBench_Width(SBW, alpha, TW, DOR, RLF, ...)
}
\arguments{
\item{SBW}{Safety Berm Width}

\item{alpha}{Clearance between Truck and Safety Berm}

\item{TW}{Width of Largest Truck on the Bench}

\item{DOR}{Dumping Radius of Loader at Maximum Height}

\item{RLF}{Radius of Level Floor}
}
\description{
Title Working Bench Design
SBW = 2*HSB , HSB  = Height of Safety Berm, HSB, Greater than or equal to (>=)  Height of Largest Truck Axle
}
