% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_item.R
\name{find_item}
\alias{find_item}
\title{Find item by name or ID/style}
\usage{
find_item(name = NULL, id = NULL, style = 0)
}
\arguments{
\item{name}{Character string with the name of a Minecraft item (specify
either \code{name} or \code{id}, not both)}

\item{id}{A numeric or character string with the ID of a Minecraft item
(specify either \code{name} or \code{id}, not both)}

\item{style}{A numeric or character string with the style of a Minecraft
item (use this argument only if querying by \code{id} is provided)}
}
\value{
Data frame with a row or set of rows from \code{\link{mc_items}} that
match the queried name, ID, and / or style.
}
\description{
Find a Minecraft item by name or ID. If querying an item by ID, the search
can also specify item style.
}
\details{
If \code{name} is provided, we first look to see whether there
is an exact match to the \code{name} column in \code{\link{mc_items}}. If there
is, we return that row. If not, we use \code{\link[=grep]{grep()}}
with \code{ignore.case=TRUE} and return matching rows.\preformatted{If instead `id` is provided, we return the row with that id.
The default is to return the row with that ID and `style==0`,
or whatever style was provided. If `style` is NULL, we return
all rows with that ID.
}
}
\examples{
find_item(name = "Oak")
find_item(id = 5)
find_item(id = 5, style = 5)

}
\seealso{
\code{\link[=find_entity]{find_entity()}}, \code{\link{mc_items}}
x
}
