% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_item.R
\name{find_item}
\alias{find_item}
\title{Find item IDs by name}
\usage{
find_item(name)
}
\arguments{
\item{name}{Name of an item}
}
\value{
Data frame defined by a row or set of rows from \code{\link{mc_items}}.
}
\description{
Find item ID by name
}
\details{
If an exact match, just return that row. If no exact match, use grep and return all matching rows.
}
