% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlock.R
\name{getBlock}
\alias{getBlock}
\title{Determine block type at some position}
\usage{
getBlock(x, y, z)
}
\arguments{
\item{x}{north/south position}

\item{y}{height}

\item{z}{east/west position}
}
\value{
None.
}
\description{
Determine block type at position (x,y,z)
}
\examples{
\dontrun{
mc_connect()
h <- getHeight(0,0)
getBlock(0,h,0)
}

}
