% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlock.R
\name{getBlock}
\alias{getBlock}
\title{Determine block type at some position}
\usage{
getBlock(con, x, y, z)
}
\arguments{
\item{con}{Socket connection to minecraft server}

\item{x}{north/south position}

\item{y}{height}

\item{z}{east/west position}
}
\value{
None.
}
\description{
Determine block type at position (x,y,z)
}
\examples{
\dontrun{
mc <- mc_connect()
h <- getHeight(mc, 0,0)
getBlock(mc,  0,h,0)
}

}
