% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlock.R
\name{getBlocks}
\alias{getBlocks}
\title{Determine block types in a cuboid}
\usage{
getBlocks(con, x0, y0, z0, x1, y1, z1)
}
\arguments{
\item{con}{Socket connection to minecraft server}

\item{x0}{north/south position}

\item{y0}{height}

\item{z0}{east/west position}

\item{x1}{north/south position}

\item{y1}{height}

\item{z1}{east/west position}
}
\value{
An array of integers of block IDs
}
\description{
Determine block types in a cuboid defined by (x0,y0,z0) to (x1,y1,z1)
}
\examples{
\dontrun{
mc <- mc_connect()
h <- getHeight(mc, 0,0)
getBlocks(mc,  0,h,0, 0, h+5, 0)
}

}
