% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlock.R
\name{getBlockWithData}
\alias{getBlockWithData}
\title{Determine block type and style at some position}
\usage{
getBlockWithData(x, y, z)
}
\arguments{
\item{x}{north/south position}

\item{y}{height}

\item{z}{east/west position}
}
\value{
None.
}
\description{
Determine block type and style at position (x,y,z)
}
\examples{
\dontrun{
mc_connect()
h <- getHeight(0,0)
getBlockWithData(0,h,0)
}

}
