% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeight.R
\name{getHeight}
\alias{getHeight}
\title{Get height of the world at a given position}
\usage{
getHeight(con, x, z)
}
\arguments{
\item{con}{Socket connection to minecraft server}

\item{x}{north/south position}

\item{z}{east/west position}
}
\value{
Integer with height of block just above the last bit of non-air.
}
\description{
Get height of the world at a given position
}
\examples{
\dontrun{
mc <- mc_connect()
getHeight(0,0)
}

}
