% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerOrientation.R
\name{getPlayerDirection}
\alias{getPlayerDirection}
\title{Get direction vector}
\usage{
getPlayerDirection()
}
\value{
A numeric vector of length 3 with coordinates of current direction
}
\description{
Get a vector describing the current direction a player is facing.
}
\examples{
\dontrun{
getDirection()
}
}
