% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerOrientation.R
\name{getPlayerDirection}
\alias{getPlayerDirection}
\title{Get player direction as unit vector}
\usage{
getPlayerDirection(player_id = NULL)
}
\arguments{
\item{player_id}{Integer giving the ID of a player. You can find IDs of all 
current players using \code{\link{getPlayerIds}}.}
}
\value{
A numeric vector of length 3 with coordinates of the player's current 
   direction as a unit vector.
}
\description{
Returns a unit vector describing the current direction a player is facing.
The default is to get the direction for the current player, but the directions
of other players can be gotten using the \code{player_id} argument.
}
\examples{
\dontrun{
getPlayerDirection()

example_playerId <- getPlayerIds()[1]
getPlayerDirection(example_playerId) 
}

}
