% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerOrientation.R
\name{getPlayerPitch}
\alias{getPlayerPitch}
\title{Get player pitch}
\usage{
getPlayerPitch(player_id = NULL)
}
\arguments{
\item{player_id}{Integer giving the ID of a player. You can find IDs of all
current players using \code{\link{getPlayerIds}}.}
}
\value{
A double vector of length one with a value between
-90 (looking straight up) and 90 (looking straight down)
giving the pitch of the player's viewpoint.
}
\description{
Return the player's pitch (angle in the up / down direction).
The default is to get the rotation for the first player spawned in the Minecraft 
world, but this can be run for a different player by using the \code{player_id}
argument.
}
\examples{
\dontrun{
getPlayerPitch()

example_playerId <- getPlayerIds()[1]
getPlayerPitch(example_playerId)
}

}
