% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerOrientation.R
\name{getPlayerPitch}
\alias{getPlayerPitch}
\title{Get player pitch}
\usage{
getPlayerPitch(player_id = NULL)
}
\arguments{
\item{player_id}{Integer giving the ID of a player. You can find IDs of all
current players using \code{\link[=getPlayerIds]{getPlayerIds()}}.}
}
\value{
A numeric value between -90 and +90, giving the pitch of
the player's viewpoint, with -90 indicating the player is
looking straight up and +90 indicating that the player is
looking straight down.
}
\description{
Return the player's pitch (angle in the up / down direction).
The default is to get the rotation for the first player spawned in the Minecraft
world, but this can be run for a different player by using the \code{player_id}
argument.
}
\examples{
\dontrun{
getPlayerPitch()

example_playerId <- getPlayerIds()[1]
getPlayerPitch(example_playerId)
}

}
\seealso{
\code{\link[=getPlayerDirection]{getPlayerDirection()}}, \code{\link[=getPlayerRotation]{getPlayerRotation()}}, \code{\link[=getPlayerPos]{getPlayerPos()}}
}
