% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerPos.R
\name{getPlayerPos}
\alias{getPlayerPos}
\title{Get player position}
\usage{
getPlayerPos(id = NULL)
}
\arguments{
\item{id}{Entity id}
}
\value{
vector (x,y,z)
}
\description{
Get entity position. The default is to get the player position, but other
positions can be gotten using the \code{id} argument.
}
\examples{
\dontrun{
mc_connect()
getPlayerPos()

example_entity <- getPlayerIds()[1]
getPlayerPos(id = example_entity)
}

}
\seealso{
[getPlayerPos]
}
