% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerOrientation.R
\name{getPlayerRotation}
\alias{getPlayerRotation}
\title{Get player rotation}
\usage{
getPlayerRotation(player_id = NULL)
}
\arguments{
\item{player_id}{Integer giving the ID of a player. You can find IDs of all
current players using \code{\link[=getPlayerIds]{getPlayerIds()}}.}
}
\value{
A numeric value between 0 and +360 indicating the direction
that the player is facing, with 0 being south (positive z), 90
being west (negative x), 180 being north (negative z), 270
being east (positive x), and 360 being back at the south again.
}
\description{
Get the current rotation of a player. The default is to get the rotation for
the first player spawned in the Minecraft world, but this can be run for a different
player by using the \code{player_id} argument.
}
\examples{
\dontrun{
getPlayerRotation()

example_playerId <- getPlayerIds()[1]
getPlayerRotation(example_playerId)
}

}
\seealso{
\code{\link[=getPlayerDirection]{getPlayerDirection()}}, \code{\link[=getPlayerPitch]{getPlayerPitch()}}, \code{\link[=getPlayerPos]{getPlayerPos()}}
}
