% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerOrientation.R
\name{getPlayerRotation}
\alias{getPlayerRotation}
\title{Get player rotation}
\usage{
getPlayerRotation(player_id = NULL)
}
\arguments{
\item{player_id}{Entity id}
}
\value{
A double vector of length 1 with a value between 
-360 and 360 giving the angle of rotation of the player.
}
\description{
Get the current rotation of the player. This can be run for a different 
player by using the \code{player_id} argument.
}
\examples{
\dontrun{
getPlayerRotation()

example_playerId <- getPlayerIds()[1]
getPlayerRotation(example_playerId)
}

}
\seealso{
\code{\link{getPlayerDirection}}
}
