% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerPos.R
\name{getPlayerTilePos}
\alias{getPlayerTilePos}
\title{Get player tile position}
\usage{
getPlayerTilePos(player_id = NULL)
}
\arguments{
\item{player_id}{Optional player ID}
}
\value{
vector (x,y,z)
}
\description{
Get player tile position
}
\examples{
\dontrun{
mc_connect()
getPlayerTilePos()

example_playerId <- getPlayerIds()[1]
getPlayerPos(example_playerId)
}

}
\seealso{
[getPlayerPos]
}
