% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{mc_connect}
\alias{mc_connect}
\title{Create connection to minecraft server}
\usage{
mc_connect(address = "localhost", port = 4711)
}
\arguments{
\item{address}{The url for the minecraft server}

\item{port}{The port to use for the connection}
}
\value{
A socket connection to the minecraft server, hopefully.
}
\description{
Create connection to minecraft server
}
