% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_items.R
\docType{data}
\name{mc_items}
\alias{mc_items}
\title{Minecraft item information}
\format{Data frame with columns `name`, `id`, and `style`.}
\source{
\url{http://minecraft-ids.grahamedgecombe.com}
}
\usage{
data(mc_items)
}
\description{
Minecraft item information
}
\examples{
data(mc_items)
mc_items[grep("Dirt", mc_items$name),]
}
\keyword{datasets}
