% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_items.R
\docType{data}
\name{mc_items}
\alias{mc_items}
\title{Minecraft item information}
\format{Data frame with columns \code{name}, \code{id}, and \code{style}.}
\source{
\url{http://minecraft-ids.grahamedgecombe.com}
}
\usage{
data(mc_items)
}
\description{
Dataset with the name, ID, and style of possible Minecraft items.
}
\examples{
data(mc_items)
mc_items[grep("Dirt", mc_items$name),]
}
\seealso{
\code{\link[=find_item]{find_item()}}, \code{\link[=find_entity]{find_entity()}}
}
\keyword{datasets}
