% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlock.R
\name{setBlock}
\alias{setBlock}
\title{Place a block}
\usage{
setBlock(x, y, z, id, style = 0)
}
\arguments{
\item{x}{A numeric string with north/south position}

\item{y}{A numeric string with height}

\item{z}{A numeric string with east/west position}

\item{id}{A numeric or character string with the ID of a Minecraft item
(specify either \code{name} or \code{id}, not both)}

\item{style}{A numeric or character string with the style of a Minecraft
item (use this argument only if querying by \code{id} is provided)}
}
\value{
None.
}
\description{
Place a block at position (x,y,z) by type id
}
\examples{
\dontrun{
mc_connect()
h <- getHeight(0, 0)
setBlock(0, h, 0, 46)
}

}
