% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBlock.R
\name{setBlocks}
\alias{setBlocks}
\title{Place blocks in a cuboid}
\usage{
setBlocks(x0, y0, z0, x1, y1, z1, block_id)
}
\arguments{
\item{x0}{north/south position}

\item{y0}{height}

\item{z0}{east/west position}

\item{x1}{north/south position}

\item{y1}{height}

\item{z1}{east/west position}

\item{block_id}{block id}
}
\value{
None.
}
\description{
Place blocks in cuboid defined by (x0,y0,z0) and (x1,y1,z1)
}
\examples{
\dontrun{
mc_connect()
h <- getHeight(0,0)
setBlock(0,h,0,  46)
}

}
