% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlayerOrientation.R
\name{setPlayerDirection}
\alias{setPlayerDirection}
\title{Set a player's direction}
\usage{
setPlayerDirection(x, y, z, id)
}
\arguments{
\item{x}{east/west direction}

\item{y}{up/down direction}

\item{z}{north/south direction}

\item{id}{Player or entity ID}
}
\description{
Set a player's direction vector (x,y,z)
}
\details{
(\code{x}, \code{y}, \code{z}) define a unit vector to which the player will now point.
}
\note{
Only works with \href{https://github.com/zhuowei/RaspberryJuice}{RaspberryJuice}
version 1.11 or later.
}
\examples{
\dontrun{
mc_connect()
getPlayerIds()
setPlayerDirection(45, 355)
}

}
\seealso{
\code{\link[=getPlayerDirection]{getPlayerDirection()}}, \code{\link[=setPlayerRotation]{setPlayerRotation()}}, \code{\link[=setPlayerPitch]{setPlayerPitch()}},
\code{\link[=setPlayerPos]{setPlayerPos()}}
}
