% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlayerPos.R
\name{setPlayerPos}
\alias{setPlayerPos}
\title{Change player position}
\usage{
setPlayerPos(x, y, z, id = NULL)
}
\arguments{
\item{x}{north/south position}

\item{y}{height}

\item{z}{east/west position}

\item{id}{Entity id}
}
\value{
None.
}
\description{
Move player to position (x,y,z). The default is to move your player, but
other entities can also be moved using the \code{id} argument.
}
\examples{
\dontrun{
mc_connect()
p <- getPlayerPos()
setPlayerPos(mc, 0, p + 5, 0)

example_entity <- getPlayerIds()[1]
getPlayerPos(id = example_entity)
setPlayerPos(0, p, 0, id = example_entity)
getPlayerPos(id = example_entity)
}

}
