% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setPlayerPos.R
\name{setPlayerTilePos}
\alias{setPlayerTilePos}
\title{Change player position to a tile}
\usage{
setPlayerTilePos(x, y, z, player_id = NULL)
}
\arguments{
\item{x}{north/south position}

\item{y}{height}

\item{z}{east/west position}
}
\value{
None.
}
\description{
Move player to tile position (x,y,z)
}
\examples{
\dontrun{
mc_connect()
p <- getPlayerTilePos()
setPlayerTilePos(0, p + 5, 0)

example_entity <- getPlayerIds()[1]
setPlayerTilePos(0, p, 0, example_entity)
}

}
