% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawnEntity.R
\name{spawnEntity}
\alias{spawnEntity}
\title{Spawn an entity}
\usage{
spawnEntity(x, y, z, typeid)
}
\arguments{
\item{x}{east/west position (east is positive)}

\item{y}{up/down position (up is positive)}

\item{z}{north/south position (south is positive)}

\item{typeid}{Integer giving the type of entity. See the
output of \code{[etEntityTypes()] or use }find_entity()` to search.}
}
\value{
A number giving the spawned entity's ID.
}
\description{
Spawn an entity of a given type at a given position.
}
\examples{
\dontrun{
mc_connect()
witch_typeid <- find_entity("witch")$id
witch_id <- spawnEntity(0, 20, 0, witch_typeid)
getPlayerName(witch_id)
}

}
\seealso{
\code{\link[=getEntityTypes]{getEntityTypes()}}, \code{\link[=find_entity]{find_entity()}}
}
